Exercise 3.01
Note that SealedSecrets are removed and secrets should be created manually
with kubectl for now. This is more straightforward than using SealedSecrets
since a new private key is generated each time the cluster is re-created and thus makes it inconvenient to have to re-encrypt the sealed secret configs each time.